合并

+++ pkg/task/tranfer.go
@@ -3,12 +3,16 @@ package task
 import (
 	"context"
 	"encoding/json"
-	model "github.com/HFO4/cloudreve/models"
-	"github.com/HFO4/cloudreve/pkg/filesystem"
-	"github.com/HFO4/cloudreve/pkg/util"
 	"os"
 	"path"
 	"path/filepath"
+	"strings"
+
+	model "github.com/cloudreve/Cloudreve/v3/models"
+	"github.com/cloudreve/Cloudreve/v3/pkg/cluster"
+	"github.com/cloudreve/Cloudreve/v3/pkg/filesystem"
+	"github.com/cloudreve/Cloudreve/v3/pkg/filesystem/fsctx"
+	"github.com/cloudreve/Cloudreve/v3/pkg/util"
 )
 
 // TransferTask 文件中转任务
@@ -23,9 +27,14 @@ type TransferTask struct {
 
 // TransferProps 中转任务属性
 type TransferProps struct {
-	Src    []string `json:"src"`    // 原始目录
-	Parent string   `json:"parent"` // 父目录
-	Dst    string   `json:"dst"`    // 目的目录ID
+	Src      []string          `json:"src"`      // 原始文件
+	SrcSizes map[string]uint64 `json:"src_size"` // 原始文件的大小信息，从机转存时使用
+	Parent   string            `json:"parent"`   // 父目录
+	Dst      string            `json:"dst"`      // 目的目录ID
+	// 将会保留原始文件的目录结构，Src 除去 Parent 开头作为最终路径
+	TrimPath bool `json:"trim_path"`
+	// 负责处理中专任务的节点ID
+	NodeID uint `json:"node_id"`
 }
 
 // Props 获取任务属性
@@ -89,7 +98,38 @@ func (job *TransferTask) Do() {
 
 	for index, file := range job.TaskProps.Src {
 		job.TaskModel.SetProgress(index)
-		err = fs.UploadFromPath(context.Background(), file, path.Join(job.TaskProps.Dst, filepath.Base(file)))
+
+		dst := path.Join(job.TaskProps.Dst, filepath.Base(file))
+		if job.TaskProps.TrimPath {
+			// 保留原始目录
+			trim := util.FormSlash(job.TaskProps.Parent)
+			src := util.FormSlash(file)
+			dst = path.Join(job.TaskProps.Dst, strings.TrimPrefix(src, trim))
+		}
+
+		if job.TaskProps.NodeID > 1 {
+			// 指定为从机中转
+
+			// 获取从机节点
+			node := cluster.Default.GetNodeByID(job.TaskProps.NodeID)
+			if node == nil {
+				job.SetErrorMsg("从机节点不可用", nil)
+			}
+
+			// 切换为从机节点处理上传
+			fs.SwitchToSlaveHandler(node)
+			err = fs.UploadFromStream(context.Background(), &fsctx.FileStream{
+				File:        nil,
+				Size:        job.TaskProps.SrcSizes[file],
+				Name:        path.Base(dst),
+				VirtualPath: path.Dir(dst),
+				Src:         file,
+			}, false)
+		} else {
+			// 主机节点中转
+			err = fs.UploadFromPath(context.Background(), file, dst, 0)
+		}
+
 		if err != nil {
 			job.SetErrorMsg("文件转存失败", err)
 		}
@@ -99,15 +139,16 @@ func (job *TransferTask) Do() {
 
 // Recycle 回收临时文件
 func (job *TransferTask) Recycle() {
-	err := os.RemoveAll(job.TaskProps.Parent)
-	if err != nil {
-		util.Log().Warning("无法删除中转临时目录[%s], %s", job.TaskProps.Parent, err)
+	if job.TaskProps.NodeID == 1 {
+		err := os.RemoveAll(job.TaskProps.Parent)
+		if err != nil {
+			util.Log().Warning("无法删除中转临时目录[%s], %s", job.TaskProps.Parent, err)
+		}
 	}
-
 }
 
 // NewTransferTask 新建中转任务
-func NewTransferTask(user uint, src []string, dst, parent string) (Job, error) {
+func NewTransferTask(user uint, src []string, dst, parent string, trim bool, node uint, sizes map[string]uint64) (Job, error) {
 	creator, err := model.GetActiveUserByID(user)
 	if err != nil {
 		return nil, err
@@ -116,9 +157,12 @@ func NewTransferTask(user uint, src []string, dst, parent string) (Job, error) {
 	newTask := &TransferTask{
 		User: &creator,
 		TaskProps: TransferProps{
-			Src:    src,
-			Parent: parent,
-			Dst:    dst,
+			Src:      src,
+			Parent:   parent,
+			Dst:      dst,
+			TrimPath: trim,
+			NodeID:   node,
+			SrcSizes: sizes,
 		},
 	}
